requirements.txt
python-telegram-bot==20.7
anthropic
faster-whisper
yt-dlp
requests
